# Improve performance of calc_connected_components