See #913 and profile therein. This PR focuses on improving the performance of `calc_connected_components` by implementing a more efficient graph search and identification of connected components.

In particular, a significant amount of time is (currently) spent in this one line:
https://github.com/lanl-ansi/PowerModels.jl/blob/8e346b48c187b2a53907dc031603482f2e6a2017/src/core/data.jl#L2538

In the case where there is a single connected component, this ends up computing the union of `N` sets, all of which are identical and contain all the buses in the network. This is because the current implementation stores the entire connected component for every bus in the system, which also comes at a O(n^2) memory cost.

I replace the previous DFS search with a more memory-efficient search that indexes each connected component with the smallest bus index in that connected component. There is a O(n log n) cost for sorting bus IDs, and then a O(E) cost for the graph search. The memory cost is O(n). It also saves time because it avoids the costly computation of a set of (large) sets. Note that I kept the last step (converting to a `Set{Set{Int}}`) to comply with the previous function definition. 

Timings of `make_basic_network` after the changes: times are improved, the impact on memory is marginal

| Case | #buses | old time (s) | new time (s) | old memory (MiB) | new memory (MiB) |
|:--|--:|--:|--:|--:|--:|
| `1354_pegase` | 1354 | 0.21 | 0.30 | 28.5 | 28.9 |
| `2869_pegase` | 2869 | 0.62 | 0.43 | 66.8 | 65.3 |
| `9241_pegase` | 9241 | 2.50 | 1.67 | 208.4 | 204.2 |
| `13659_pegase` | 13659 | 5.72 | 2.47 | 303.1 | 306.2 |

and corresponding profile:
![prof_new](https://github.com/lanl-ansi/PowerModels.jl/assets/9593025/bd1c492d-089e-471e-82ed-f7b8cfe0e573)