@mtanneau 2024-06-11
@ccoffrin I believe we can also remove the [`_cc_dfs` function](https://github.com/lanl-ansi/PowerModels.jl/blob/8e346b48c187b2a53907dc031603482f2e6a2017/src/core/data.jl#L2544-L2560). I didn't because I was afraid it might break something downstream.

Side note: Graphs.jl has a lot of functionality to do these kind of graph operations, which are likely more efficient (at the cost of yet another dependency).